requirements.txt
amqp==2.6.0
asgiref==3.2.10
attrs==19.3.0
beautifulsoup4==4.9.1
billiard==3.6.3.0
bs4==0.0.1
celery==4.4.6
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
django-compat==1.0.15
djangorestframework==3.11.0
future==0.18.2
idna==2.10
kombu==4.6.11
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.9.0
pyparsing==2.4.7
pytest==5.4.3
pytest-dependency==0.5.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
wcwidth==0.2.5
